requirements.txt
Flask==3.0.0
openai==1.3.4
python-dotenv==1.0.0
